# contract/openapi.yaml
openapi: 3.1.0
info:
  title: КОСМОБЕРЕГ API
  version: 1.0.0
  description: |
    Контракт между фронтендом и бэкендом платформы КОСМОБЕРЕГ.

    ## Как этим пользоваться

    Это **единственный источник истины** об интерфейсе между фронтом и бэком.
    Фронт кодируется против контракта (с моками, пока бэка нет), бэк реализует
    контракт. Стыковка = смена базового URL.

    Изменение контракта — только по согласованию обеих сторон. Правило простое:
    **сначала правим contract/openapi.yaml, потом код.**

    ## Ключевое решение: локализация на клиенте

    API **никогда не возвращает переведённый текст**. Возвращаются доменные коды
    (`status: "problem"`, `role: "observer"`) и ключи локализации
    (`name_key: "seg.ustie.name"`). Клиент переводит их через `i18n.js`.

    Причины:
    - фронт уже несёт словарь на 3 языка (ru/en/zh) — дублировать его на бэке не нужно;
    - смена языка не требует запроса к серверу;
    - добавление 4-го языка не затрагивает бэкенд вообще.

    Исключение — **пользовательский контент** (описания в отчётах, имена
    пользователей): возвращается как есть, без перевода.

    ## Что вычисляется только на сервере

    Клиент не может это подделать, поэтому такие значения приходят с бэка и
    никогда не принимаются от фронта:

    - `attention_index` и факторы `s`, `c`, `t`, `a`;
    - переход роли `student → observer` (только после завершения курса);
    - достижение консенсуса по разметке;
...
    - смена статуса участка.
servers:
...
security:
  - bearerAuth: []
paths:
...
  /segments/{segment_id}/imagery-requests:
    post:
      tags: [imagery]
      summary: Запросить новую съёмку участка
      description: |
        Делает связь с данными ДЗЗ двусторонней: платформа не только
        потребляет снимки, но и запрашивает их. Геопортал Роскосмоса
        принимает такие заявки штатно.

        Если аномалия на участке растёт, приоритет автоматически повышается
        до `urgent`: пока идёт обычная очередь оператора, свалка успевает
        вырасти ещё.

        Повторная заявка при активной блокируется — две заявки на одну
        съёмку означают лишние расходы и путаницу в очереди оператора.
      parameters:
        - $ref: '#/components/parameters/SegmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImageryRequestCreate' }
      responses:
        '201':
          description: Заявка отправлена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImageryRequestOut' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: По участку уже есть активная заявка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /imagery-requests:
    get:
      tags: [imagery]
      summary: Заявки на съёмку
      description: Инспектор видит только заявки своей территории.
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ImageryRequestStatus' }
        - name: segment_id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Список заявок
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ImageryRequestOut' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ==================== Медиа ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SegmentId:
      name: segment_id
      in: path
      required: true
      schema: { type: string, example: ustie }
...
  responses:
...
    Forbidden:
      description: Роль не позволяет выполнить операцию
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message_key]
      properties:
        code:
          type: string
          example: consent_required
          description: Машинный код ошибки
        message_key:
          type: string
          example: err.consent_required
          description: Ключ локализации для i18n.js — текст берётся из словаря фронта
        details:
          type: object
          additionalProperties: true
...
    ImageryRequestStatus:
      type: string
      enum: [queued, accepted, delivered, rejected]
      description: |
        Отражает реальный процесс оператора ДЗЗ:
        * `queued` — отправлена оператору
        * `accepted` — принята к съёмке
        * `delivered` — снимок получен и нарезан
        * `rejected` — отказ (облачность, приоритет других задач, ограничения)
    ImageryPriority:
      type: string
      enum: [normal, urgent]
      description: |
        `urgent` выставляется автоматически, если аномалия на участке растёт —
        пока идёт обычная очередь оператора, свалка успевает вырасти ещё.
    ImageryRequestCreate:
      type: object
      properties:
        priority: { $ref: '#/components/schemas/ImageryPriority' }
        comment:  { type: string, maxLength: 1000 }
    ImageryRequestOut:
      type: object
      required: [id, segment_id, oopt_id, priority, status, created_at]
      properties:
        id:               { type: string, format: uuid }
        segment_id:       { type: string }
        segment_name_key: { type: string, nullable: true }
        oopt_id:          { type: string }
        priority: { $ref: '#/components/schemas/ImageryPriority' }
        status:   { $ref: '#/components/schemas/ImageryRequestStatus' }
        comment:  { type: string, nullable: true }
        expected_at:
          type: string
          format: date
          nullable: true
          description: |
            Ожидаемая дата съёмки со слов оператора. Показывайте её:
            без неё заявка выглядит отправленной в пустоту.
        reject_reason: { type: string, nullable: true }
        external_id:
          type: string
          nullable: true
          description: Идентификатор заявки на стороне оператора ДЗЗ — для сверки
        delivered_at: { type: string, format: date-time, nullable: true }
        created_at:   { type: string, format: date-time }